сделана гавно структура проекта полное ващееее кал отстой ебанный просто
