+++ tokens.py
@@ -1,0 +1,2 @@
+api_hash = 111
+api_id = 11
